apply asyncfeature on yamnet ios
