+++ yamnet/ZeticMLangeYAMNET-iOS/ZeticMLange.framework/_CodeSignature/CodeResources
@@ -6,31 +6,31 @@
 	<dict>
 		<key>Info.plist</key>
 		<data>
-		wErxS/zsQb2V03coMoqiZVli/tw=
+		l9sxFkguH+6N2IhPSbwXrgeuFro=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/Project/arm64-apple-ios.swiftsourceinfo</key>
 		<data>
-		pwyBTMkhwrQ/vvDvE17+sr8pHvQ=
+		2BBPsdyARoMIIRjZAQianN7spHI=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.abi.json</key>
 		<data>
-		s9mKx7n1FFmL8SHyb9LAfuyfV4w=
+		42MF3M9bhzKPN/3xJkX0EMnkCAE=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.private.swiftinterface</key>
 		<data>
-		brzubGdNtN6Xxh9YVgvBTY6E29I=
+		PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftdoc</key>
 		<data>
-		bzNz6HGWMcHm65ogxnmrxB923hQ=
+		48LzpYNIYTiZbgdXeQaHNYK0aXs=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftinterface</key>
 		<data>
-		brzubGdNtN6Xxh9YVgvBTY6E29I=
+		PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftmodule</key>
 		<data>
-		C/AgYVEOw/oBJFJrktOx/zbV01c=
+		V9w0JZdwB0/aQNlGQH8wC9NCzXA=
 		</data>
 	</dict>
 	<key>files2</key>
@@ -39,66 +39,66 @@
 		<dict>
 			<key>hash</key>
 			<data>
-			pwyBTMkhwrQ/vvDvE17+sr8pHvQ=
+			2BBPsdyARoMIIRjZAQianN7spHI=
 			</data>
 			<key>hash2</key>
 			<data>
-			Smn4R8gETupaRUpI6qSW5MJw6GpO+E884IxNMSS0DzY=
+			3CY6Za8hZQyCiNMdibJwFWiNl0nI6uDlLEadFqphnFQ=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.abi.json</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			s9mKx7n1FFmL8SHyb9LAfuyfV4w=
+			42MF3M9bhzKPN/3xJkX0EMnkCAE=
 			</data>
 			<key>hash2</key>
 			<data>
-			GslrWi3oh8YKT3JU0I3JZj5Y5aLPb6/C87DvNQOC17A=
+			NhfyCiQCxMBsD1LUx+3cgDE5l5A3rGp4EHS1QQqloeI=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.private.swiftinterface</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			brzubGdNtN6Xxh9YVgvBTY6E29I=
+			PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
 			</data>
 			<key>hash2</key>
 			<data>
-			Vku8/kXCJ+IMCvIHKeELCWYf3IFL0IQSs7y0QU5lIIA=
+			eUJF0F4YispGrblZ0O8QIJ+KMj5PoL1IHsvTaA7rU/w=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftdoc</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			bzNz6HGWMcHm65ogxnmrxB923hQ=
+			48LzpYNIYTiZbgdXeQaHNYK0aXs=
 			</data>
 			<key>hash2</key>
 			<data>
-			+ZStWnx5wrkGhMkP90UsyJBiAWto+ySmSVLo3eFKl2s=
+			nnKIs83YS50RLRDOhnU3MdBSUz5tfKr7SY0ANK2Z90c=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftinterface</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			brzubGdNtN6Xxh9YVgvBTY6E29I=
+			PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
 			</data>
 			<key>hash2</key>
 			<data>
-			Vku8/kXCJ+IMCvIHKeELCWYf3IFL0IQSs7y0QU5lIIA=
+			eUJF0F4YispGrblZ0O8QIJ+KMj5PoL1IHsvTaA7rU/w=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftmodule</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			C/AgYVEOw/oBJFJrktOx/zbV01c=
+			V9w0JZdwB0/aQNlGQH8wC9NCzXA=
 			</data>
 			<key>hash2</key>
 			<data>
-			ZWFiBzAz/T/fFSTanOgdXgeD6uGBBLMKUD5sSRIn9bw=
+			DGndV2QAFryCzaH7TpRk4nezLhiJyhxfvFnjFihRm8s=
 			</data>
 		</dict>
 	</dict>
