+++ yamnet/ZeticMLangeYAMNET-iOS/ZeticMLangeYAMNET-iOS/ZeticMLangeYAMNET_iOSApp.swift
@@ -4,7 +4,7 @@ import SwiftUI
 struct ZeticMLangeYAMNet_iOSApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            YAMNetView()
         }
     }
 }
